Convert to VersionedNodeType pattern for v2 migration
Move existing Pipedrive node into v1/ directory and introduce a VersionedNodeType
entry point, following the established Discord node pattern. This is a pure
structural refactor with zero behavior change — existing workflows continue
working unchanged.

- Move GenericFunctions.ts to v1/GenericFunctions.ts
- Move Pipedrive.node.ts to v1/PipedriveV1.node.ts with constructor pattern
- Create new Pipedrive.node.ts as VersionedNodeType entry point (defaultVersion: 1)
- Update PipedriveTrigger and test imports to reference v1/ path
- Add oldVersionNotice to v1 description properties
